Review comments, cleanup, componentization
